+++ tiger-rbel/src/main/java/de/gematik/rbellogger/util/RbelJexlExecutor.java
@@ -8,6 +8,7 @@ import com.google.common.base.CharMatcher;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.test.tiger.common.TokenSubstituteHelper;
 import de.gematik.test.tiger.common.config.TigerConfigurationLoader;
+import de.gematik.test.tiger.common.exceptions.TigerJexlException;
 import de.gematik.test.tiger.common.jexl.TigerJexlContext;
 import de.gematik.test.tiger.common.jexl.TigerJexlExecutor;
 import java.util.*;
@@ -78,8 +79,25 @@ public class RbelJexlExecutor {
     }
   }
 
+  /**
+   * This method expands one expression to a list of all possible concrete-value interpretations.
+   * Every containing RbelPath-expression is resolved to ALL POSSIBLE concrete values (there can be
+   * more then one). Since the resulting expressions are static there will be a value added for
+   * every result. For two RbelPath-expressions referencing three different elements each the
+   * resulting list will contain all six combinations.
+   *
+   * @param jexlExpression The expression to be analyzed
+   * @param mapContext The map-context to be used
+   * @return A list of all concrete JEXL-expressions
+   */
   private static List<String> evaluateRbelPathExpressions(
       String jexlExpression, TigerJexlContext mapContext) {
+    // always skip the evaluation if there is no element in the current context
+    // necessary hack, remove with AST
+    if (mapContext.getRootElement() == null) {
+      return List.of(jexlExpression);
+    }
+
     List<String> resultingPaths = List.of(jexlExpression);
     for (var potentialPath : extractPotentialRbelPaths(jexlExpression)) {
       if (!(potentialPath.startsWith("$.") || potentialPath.startsWith("@."))) {
@@ -88,7 +106,7 @@ public class RbelJexlExecutor {
       List<String> previousIterationPaths = new ArrayList<>(resultingPaths);
       List<String> newPaths = new ArrayList<>();
       for (var expression : previousIterationPaths) {
-        evaluateAndCollectSubPaths(mapContext, potentialPath, expression, newPaths);
+        evaluatePathsAndCollectAllResults(mapContext, potentialPath, expression, newPaths);
       }
       if (!newPaths.isEmpty()) {
         resultingPaths = newPaths;
@@ -97,22 +115,29 @@ public class RbelJexlExecutor {
     return resultingPaths;
   }
 
-  private static void evaluateAndCollectSubPaths(
+  private static void evaluatePathsAndCollectAllResults(
       TigerJexlContext mapContext, String potentialPath, String expression, List<String> newPaths) {
     final List<String> pathResults =
-        extractPathAndConvertToString(
+        new ArrayList<>(extractPathAndConvertToString(
             potentialPath.startsWith("@.")
                 ? mapContext.getCurrentElement()
                 : mapContext.getRootElement(),
             potentialPath.startsWith("@.")
                 ? potentialPath.replaceFirst("@\\.", "\\$.")
-                : potentialPath);
+                : potentialPath));
     if (pathResults.isEmpty()
         || pathResults.stream().anyMatch(s -> !CharMatcher.ascii().matchesAllOf(s))) {
-      return;
+      if (mapContext.shouldIgnoreEmptyRbelPaths()) {
+        pathResults.add(null);
+      } else {
+        throw new TigerJexlException(
+            "Error while pre-processing JEXL-expression: No values found for RbelPath '"
+                + potentialPath
+                + "'!");
+      }
     }
     for (String pathResult : pathResults) {
-      final String id = "replacedPath_" + RandomStringUtils.randomAlphabetic(20); // NOSONAR
+      final String id = "replacedPath_" + RandomStringUtils.randomAlphabetic(20).toLowerCase(); // NOSONAR
       mapContext.put(id, pathResult);
       newPaths.add(expression.replace(potentialPath, id));
     }